+++ test/unit/ui/camera/fly_to.test.ts
@@ -986,5 +986,14 @@ describe('camera', () => {
             // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
             expect(camera.getPadding()).toEqual({top: 300, bottom: 0, left: 0, right: 0});
         });
+
+        test('preloadOnly does not change padding', () => {
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+            const camera = createCamera();
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+            camera.flyTo({center: [10, 10], padding: {top: 200}, preloadOnly: true});
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+            expect(camera.getPadding()).toEqual({top: 0, bottom: 0, left: 0, right: 0});
+        });
     });
 });